requirements.txt
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
django-filter==24.1
psycopg2-binary==2.9.9
django-redis==5.4.0
gunicorn==22.0.0
python-decouple==3.8
dj-database-url==2.1.0
Pillow==10.3.0
celery==5.3.6
redis==5.0.1
